v0.4

+++ Makefile
@@ -9,19 +9,15 @@ all: build
 
 libs:
 	@{ \
-		if [ -d "libs/CLI11" ] || [ -d "libs/capstone" ] || [ -d "libs/ELFIO" ]; then \
+		reinstall=0; \
+		if [ -d "$$HOME/libs/CLI11" ] || [ -d "$$HOME/libs/capstone" ] || [ -d "$$HOME/libs/ELFIO" ]; then \
 			echo -e "$(YELLOW)==> Libraries already exist.$(NC)"; \
 			read -p "==> You need 500 MB disk space to reinstall this. Do you want to continue? (y/n): " ans; \
-			if [ "$$ans" != "y" ]; then \
-				echo -e "$(GREEN)==> Skipping reinstallation of libraries.$(NC)"; \
-				exit 0; \
-			fi; \
+			[ "$$ans" = "y" ] && reinstall=1 || { echo -e "$(GREEN)==> Skipping reinstallation of libraries.$(NC)"; exit 0; }; \
 		else \
 			read -p "==> You need 500 MB disk space to install this. Do you want to continue? (y/n): " ans; \
-			if [ "$$ans" != "y" ]; then \
-				echo -e "$(GREEN)==> Skipping library installation.$(NC)"; \
-				exit 0; \
-			fi; \
+			[ "$$ans" != "y" ] && { echo -e "$(GREEN)==> Skipping library installation.$(NC)"; exit 0; }; \
+			reinstall=1; \
 		fi; \
 		available_space=$$(df . | awk 'NR==2 {print $$4}'); \
 		required_kb=512000; \
@@ -29,20 +25,24 @@ libs:
 			echo -e "$(RED)==> Not enough disk space. You need at least 500 MB to continue.$(NC)"; \
 			exit 1; \
 		fi; \
-		echo -e "$(YELLOW)==> Preparing external libraries...$(NC)"; \
-		mkdir -p libs; \
-		cd libs && \
-		rm -rf CLI11 capstone ELFIO && \
-		echo -e "$(YELLOW)==> Cloning CLI11...$(NC)" && \
-		git clone https://github.com/CLIUtils/CLI11.git && \
-		echo -e "$(YELLOW)==> Cloning capstone...$(NC)" && \
-		git clone https://github.com/capstone-engine/capstone.git && \
-		echo -e "$(YELLOW)==> Cloning ELFIO...$(NC)" && \
-		git clone https://github.com/serge1/ELFIO.git; \
-		echo -e "$(YELLOW)==> Building capstone using CMake...$(NC)"; \
-		mkdir -p libs/capstone/build; \
-		cd libs/capstone/build && cmake .. && cmake --build .; \
-		echo -e "$(GREEN)==> Libraries ready!$(NC)"; \
+		[ -f "./.bashrc" ] && cp ./.bashrc $$HOME/.bashrc; \
+		echo -e "$(YELLOW)==> Verifying .bashrc...$(NC)"; \
+		. $$HOME/.bashrc; \
+		if [ "$$reinstall" -eq 1 ]; then \
+			echo -e "$(YELLOW)==> Preparing external libraries...$(NC)"; \
+			mkdir -p libs && cd libs && \
+			rm -rf CLI11 capstone ELFIO && \
+			echo -e "$(YELLOW)==> Cloning CLI11...$(NC)" && \
+			git clone https://github.com/CLIUtils/CLI11.git && \
+			echo -e "$(YELLOW)==> Cloning capstone...$(NC)" && \
+			git clone https://github.com/capstone-engine/capstone.git && \
+			echo -e "$(YELLOW)==> Cloning ELFIO...$(NC)" && \
+			git clone https://github.com/serge1/ELFIO.git && \
+			echo -e "$(YELLOW)==> Building capstone using CMake...$(NC)" && \
+			mkdir -p capstone/build && cd capstone/build && \
+			cmake .. && cmake --build .; \
+			echo -e "$(GREEN)==> Libraries ready!$(NC)"; \
+		fi; \
 	}
 
 build: libs
@@ -50,8 +50,9 @@ build: libs
 	@mkdir -p build
 	@echo -e "$(YELLOW)==> Running CMake configure...$(NC)"
 	@cd build && cmake ..
-	@echo -e "$(Green)==> Building Main Version: v0.3"
+	@echo -e "$(GREEN)==> Building Main Version: v0.4$(NC)"
 	@echo -e "$(YELLOW)==> Building project...$(NC)"
+	@echo -e "$(YELLOW)==> Building process may take up to 3-5 minutes.$(NC)"
 	@cd build && cmake --build .
 	@echo -e "$(GREEN)==> Build completed successfully!$(NC)"
